backend/requirements.txt
fastapi
uvicorn
python-dotenv
langchain-google-genai
langchain-openai
langchain-community
langchain-core
langchain
python-jose[cryptography]
pymongo
bcrypt
fastapi-mail
pydantic
python-multipart
PyMuPDF
motor
pinecone-client
fastapi[standard]
passlib                                                                                                                             
qdrant_client
langchain_qdrant
pypdf
openai